Add `src/admin` and `src/user` dirs to distinguish them
