+++ ui/src/user/Attendance.tsx
@@ -9,9 +9,9 @@ import {
   GridToolbarQuickFilter,
 } from '@mui/x-data-grid';
 import * as XLSX from 'xlsx';
-import { useHeader } from './Layout';
-import { getHalfYearAttendances } from './client/http/default';
-import { toYYslashMMslashDDspaceHHcolonMM } from './common/date';
+import { useHeader } from '../Layout';
+import { getHalfYearAttendances } from '../client/http/default';
+import { toYYslashMMslashDDspaceHHcolonMM } from '../common/date';
 
 type Attendance = {
   id: string;
