+++ ui/src/user/MyPage.tsx
@@ -16,13 +16,13 @@ import {
   TableBody,
   Switch,
 } from '@mui/material';
-import { useHeader } from './Layout';
-import { useAuth } from './auth';
-import { Role } from './auth/role';
-import { Attendance, User } from './client/http/data';
-import { getUser, getUserAttendances, getUserAuth } from './client/http/default';
-import { toYYslashMMslashDDspaceHHcolonMM } from './common/date';
-import { useAdminMode } from './mode';
+import { useHeader } from '../Layout';
+import { useAuth } from '../auth';
+import { Role } from '../auth/role';
+import { Attendance, User } from '../client/http/data';
+import { getUser, getUserAttendances, getUserAuth } from '../client/http/default';
+import { toYYslashMMslashDDspaceHHcolonMM } from '../common/date';
+import { useAdminMode } from '../mode';
 
 const MyPage = () => {
   useHeader({ newTitle: 'Me' });
