+++ ui/src/user/SessionDetail/admin/AdminSessionDetail.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { Container, Box, Button, CircularProgress, Paper, Typography } from '@mui/material';
-import { useHeader } from '../../Layout';
-import { adminCreateSessionForm, adminDeleteSession, adminGetSession } from '../../client/http/admin';
-import { AdminSession, SessionAttendanceAppliedBy } from '../../client/http/data';
-import useHandleError from '../../common/error';
+import { useHeader } from '../../../Layout';
+import { adminCreateSessionForm, adminDeleteSession, adminGetSession } from '../../../client/http/admin';
+import { AdminSession, SessionAttendanceAppliedBy } from '../../../client/http/data';
+import useHandleError from '../../../common/error';
 import SessionInfo from '../common/SessionInfo';
 import AttendanceQrPanel from './AttendanceQrPanel';
 import SessionAttendanceTable from './AttendanceTable';
