+++ ui/src/user/SessionDetail/admin/index.ts
@@ -1,3 +1,4 @@
 import AdminSessionDetail from './AdminSessionDetail';
 
+// TODO(#57): Move this to src/admin directory.
 export default AdminSessionDetail;
