+++ ui/src/user/SessionDetail/common/UserAttendance.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { ArrowDownward, ArrowUpward } from '@mui/icons-material';
 import { CircularProgress, TableContainer, Table, TableHead, TableRow, TableCell, TableBody, Box } from '@mui/material';
-import { Attendance } from '../../client/http/data';
-import { toYYslashMMslashDDspaceHHcolonMMcolonSS } from '../../common/date';
+import { Attendance } from '../../../client/http/data';
+import { toYYslashMMslashDDspaceHHcolonMMcolonSS } from '../../../common/date';
 
 type OrderBy = 'asc' | 'desc';
