+++ ui/src/user/SessionDetail/index.ts
