+++ ui/src/user/SessionDetail/user/UserSessionDetail.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { Container, Box, Button, CircularProgress } from '@mui/material';
-import { useHeader } from '../../Layout';
-import { Session } from '../../client/http/data';
-import { getSession } from '../../client/http/default';
-import useHandleError from '../../common/error';
+import { useHeader } from '../../../Layout';
+import { Session } from '../../../client/http/data';
+import { getSession } from '../../../client/http/default';
+import useHandleError from '../../../common/error';
 import SessionInfo from '../common/SessionInfo';
 import SessionAttendanceTable from './AttendanceTable';
