+++ ui/src/user/SessionDetail/user/index.ts
